# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "AI Mafia Backend",
        "description": "Backend для игры в мафию с AI игроками",
        "version": "1.0.0"
    },
    "paths": {
        "/api/rooms/": {
            "get": {
                "tags": [
                    "rooms"
                ],
                "summary": "List Rooms",
                "description": "Получить список активных лобби для браузера комнат.\nВозвращает только комнаты со статусом 'lobby' или 'starting'.",
                "operationId": "list_rooms_api_rooms__get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Room"
                                    },
                                    "title": "Response List Rooms Api Rooms  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "rooms"
                ],
                "summary": "Create Room",
                "description": "Создать новую комнату.",
                "operationId": "create_room_api_rooms__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoomCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "RoleConfig": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "count": {
                        "type": "integer",
                        "title": "Count"
                    },
                    "canBeHuman": {
                        "type": "boolean",
                        "title": "Canbehuman",
                        "default": true
                    },
                    "canBeAI": {
                        "type": "boolean",
                        "title": "Canbeai",
                        "default": true
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "count"
                ],
                "title": "RoleConfig"
            },
            "Room": {
                "properties": {
                    "room_id": {
                        "type": "string",
                        "title": "Room Id"
                    },
                    "host_token": {
                        "type": "string",
                        "title": "Host Token"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "totalPlayers": {
                        "type": "integer",
                        "title": "Totalplayers",
                        "default": 8
                    },
                    "aiCount": {
                        "type": "integer",
                        "title": "Aicount",
                        "default": 3
                    },
                    "peopleCount": {
                        "type": "integer",
                        "title": "Peoplecount",
                        "default": 5
                    },
                    "roles": {
                        "anyOf": [
                            {
                                "additionalProperties": {
                                    "$ref": "#/components/schemas/RoleConfig"
                                },
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Roles"
                    },
                    "current_players": {
                        "type": "integer",
                        "title": "Current Players",
                        "default": 0
                    },
                    "ai_players": {
                        "type": "integer",
                        "title": "Ai Players",
                        "default": 0
                    },
                    "human_players": {
                        "type": "integer",
                        "title": "Human Players",
                        "default": 0
                    },
                    "settings": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Settings"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "room_id",
                    "host_token",
                    "status",
                    "id",
                    "created_at"
                ],
                "title": "Room"
            },
            "RoomCreate": {
                "properties": {
                    "room_id": {
                        "type": "string",
                        "title": "Room Id"
                    },
                    "host_token": {
                        "type": "string",
                        "title": "Host Token"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "totalPlayers": {
                        "type": "integer",
                        "title": "Totalplayers",
                        "default": 8
                    },
                    "aiCount": {
                        "type": "integer",
                        "title": "Aicount",
                        "default": 3
                    },
                    "peopleCount": {
                        "type": "integer",
                        "title": "Peoplecount",
                        "default": 5
                    },
                    "roles": {
                        "anyOf": [
                            {
                                "additionalProperties": {
                                    "$ref": "#/components/schemas/RoleConfig"
                                },
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Roles"
                    },
                    "current_players": {
                        "type": "integer",
                        "title": "Current Players",
                        "default": 0
                    },
                    "ai_players": {
                        "type": "integer",
                        "title": "Ai Players",
                        "default": 0
                    },
                    "human_players": {
                        "type": "integer",
                        "title": "Human Players",
                        "default": 0
                    },
                    "settings": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Settings"
                    }
                },
                "type": "object",
                "required": [
                    "room_id",
                    "host_token",
                    "status"
                ],
                "title": "RoomCreate"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}